LPFE-291: Removed ng1 module and moved index.ts to app.ng1/app.module.ts
