+++ src/app.ng1/app.module.ts
@@ -1,4 +1,4 @@
-/// <reference path='../typings/index.d.ts' />
+/// <reference path='../../typings/index.d.ts' />
 import * as angular from 'angular';
 
 // angular 1 vendors
@@ -7,53 +7,53 @@ import 'angular-local-storage';
 import 'angular-material/angular-material.css';
 
 // angular 1 app container components
-import { BasketComponent } from './app.ng1/containers/basket/basket.component';
-import { ToolbarComponent as BasketToolbarComponent } from './app.ng1/containers/basket/toolbar.component';
-import { WeekMenuComponent } from './app.ng1/containers/week-menu/week-menu.component';
-import { ToolbarComponent as WeekMenuToolbarComponent } from './app.ng1/containers/week-menu/toolbar.component';
-import { MyOrdersComponent } from './app.ng1/containers/my-orders/my-orders.component';
-import { ToolbarComponent as MyOrdersToolbarComponent } from './app.ng1/containers/my-orders/toolbar.component';
-import { PaymentComponent } from './app.ng1/containers/payment/payment.component';
-import { ToolbarComponent as PaymentToolbarComponent } from './app.ng1/containers/payment/toolbar.component';
-import { PricesComponent } from './app.ng1/containers/prices/prices.component';
-import { ToolbarComponent as PricesToolbarComponent } from './app.ng1/containers/prices/toolbar.component';
+import { BasketComponent } from './containers/basket/basket.component';
+import { ToolbarComponent as BasketToolbarComponent } from './containers/basket/toolbar.component';
+import { WeekMenuComponent } from './containers/week-menu/week-menu.component';
+import { ToolbarComponent as WeekMenuToolbarComponent } from './containers/week-menu/toolbar.component';
+import { MyOrdersComponent } from './containers/my-orders/my-orders.component';
+import { ToolbarComponent as MyOrdersToolbarComponent } from './containers/my-orders/toolbar.component';
+import { PaymentComponent } from './containers/payment/payment.component';
+import { ToolbarComponent as PaymentToolbarComponent } from './containers/payment/toolbar.component';
+import { PricesComponent } from './containers/prices/prices.component';
+import { ToolbarComponent as PricesToolbarComponent } from './containers/prices/toolbar.component';
 
 // angular 1 app presentational components
-import { MenuComponent } from './app.ng1/components/menu/menu.component';
-import { ViewMenuComponent } from './app.ng1/components/view-menu/view-menu.component';
-import { LineItemComponent } from './app.ng1/components/line-item/line-item.component';
-import { SelectorComponent } from './app.ng1/components/selector/selector.component';
-import { SizeSelectorComponent } from './app.ng1/components/size-selector/size-selector.component';
-import { QuantitySelectorComponent } from './app.ng1/components/quantity-selector/quantity-selector.component';
-import { BasketOrderComponent } from './app.ng1/components/basket-order/basket-order.component';
-import { MyOrdersItemComponent } from './app.ng1/components/my-orders-item/my-orders-item.component';
-import { PaymentStatusComponent } from './app.ng1/components/payment-status/payment-status.component';
-import { DateRangeSelectorComponent } from './app.ng1/components/date-range-selector/date-range-selector.component';
-import { UserCardComponent } from './app.ng1/components/user-card/user-card.component';
-import { SwitchComponent } from './app.ng1/components/switch/switch.component';
-import { FloorSelectorComponent } from './app.ng1/components/floor-selector/floor-selector.component';
-import { SizeLabelComponent } from './app.ng1/components/size-label/size-label.component';
-import { SidebarComponent } from './app.ng1/components/sidebar/sidebar.component';
-import { ListComponent } from './app.ng1/components/list/list.component';
+import { MenuComponent } from './components/menu/menu.component';
+import { ViewMenuComponent } from './components/view-menu/view-menu.component';
+import { LineItemComponent } from './components/line-item/line-item.component';
+import { SelectorComponent } from './components/selector/selector.component';
+import { SizeSelectorComponent } from './components/size-selector/size-selector.component';
+import { QuantitySelectorComponent } from './components/quantity-selector/quantity-selector.component';
+import { BasketOrderComponent } from './components/basket-order/basket-order.component';
+import { MyOrdersItemComponent } from './components/my-orders-item/my-orders-item.component';
+import { PaymentStatusComponent } from './components/payment-status/payment-status.component';
+import { DateRangeSelectorComponent } from './components/date-range-selector/date-range-selector.component';
+import { UserCardComponent } from './components/user-card/user-card.component';
+import { SwitchComponent } from './components/switch/switch.component';
+import { FloorSelectorComponent } from './components/floor-selector/floor-selector.component';
+import { SizeLabelComponent } from './components/size-label/size-label.component';
+import { SidebarComponent } from './components/sidebar/sidebar.component';
+import { ListComponent } from './components/list/list.component';
 
 // angular 1 app material wrapper components to be used in angular 2
-import { MdButtonComponent } from './app.ng1/components/md-button/md-button.component';
-import { MdContentComponent } from './app.ng1/components/md-content/md-content.component';
-import { MdToolbarComponent } from './app.ng1/components/md-toolbar/md-toolbar.component';
-import { MdSidenavComponent } from './app.ng1/components/md-sidenav/md-sidenav.component';
-import { MdIconComponent } from './app.ng1/components/md-icon/md-icon.component';
+import { MdButtonComponent } from './components/md-button/md-button.component';
+import { MdContentComponent } from './components/md-content/md-content.component';
+import { MdToolbarComponent } from './components/md-toolbar/md-toolbar.component';
+import { MdSidenavComponent } from './components/md-sidenav/md-sidenav.component';
+import { MdIconComponent } from './components/md-icon/md-icon.component';
 
 // angular 1 app services
-import { MenuService } from './app.ng1/models/menu';
-import { OrderService } from './app.ng1/models/order';
-import { UserService } from './app.ng1/models/user';
-import { ToastService } from './app.ng1/models/toast';
-import { BasketService } from './app.ng1/models/basket';
-import { LineItemService } from './app.ng1/models/line-item';
-import { PriceService } from './app.ng1/models/price';
+import { MenuService } from './models/menu';
+import { OrderService } from './models/order';
+import { UserService } from './models/user';
+import { ToastService } from './models/toast';
+import { BasketService } from './models/basket';
+import { LineItemService } from './models/line-item';
+import { PriceService } from './models/price';
 
 // angular 1 app filters
-import { DateFilter } from './app.ng1/filters/date.filter';
+import { DateFilter } from './filters/date.filter';
 
 // angular 1 app configs
 import { localeConfig, localStorageConfig, dateRangeSelectorConfig } from './config';
@@ -77,16 +77,16 @@ import '@angular/router';
 import { downgradeComponent, downgradeInjectable } from '@angular/upgrade/static';
 
 // angular 2 app
-import { AppComponent } from './app/app.component';
-import { RouterWrapper } from './app/ng1';
-import { ConfigService } from './app/config';
+import { AppComponent } from '../app/app.component';
+import { RouterWrapper } from '../app/core';
+import { ConfigService } from '../app/config';
 
 import {
   ExampleComponent,
   FlashMessageComponent,
   PastDaysSwitcherComponent,
   MenuCoverComponent
-} from './app/shared/components';
+} from '../app/shared/components';
 
 // -------------------------------------------------------------------------- //
 angular
@@ -178,4 +178,3 @@ angular
   }) as angular.IDirectiveFactory)
   ;
 
-import './app/main';
