+++ src/app.ng1/components/date-range-selector/date-range-selector.component.ts
@@ -2,7 +2,7 @@ import * as moment from 'moment';
 import { cloneDeep } from 'lodash';
 import { IComponentOptions, IScope, IOnChangesObject } from 'angular';
 
-import { SHORT_DATE_FORMAT } from '../../../config';
+import { SHORT_DATE_FORMAT } from '../../config';
 
 // exported types --------------------------------------------------------------
 export interface IDateRange {
