+++ src/app.ng1/components/sidebar/sidebar.component.ts
@@ -1,5 +1,5 @@
 import { IComponentOptions } from 'angular';
-import { RouterWrapper } from '../../../app/ng1';
+import { RouterWrapper } from '../../../app/core';
 
 import { IUser, UserService } from '../../models/user';
