+++ src/app.ng1/config.ts
@@ -1,4 +1,4 @@
-/// <reference path="../typings/index.d.ts" />
+/// <reference path="../../typings/index.d.ts" />
 
 import 'moment/locale/ru';
 import * as moment from 'moment';
