+++ src/app.ng1/index.scss
@@ -1,5 +1,5 @@
-@import './app.ng1/components/payment-status/payment-status';
-@import './app.ng1/components/size-label/size-label';
+@import './components/payment-status/payment-status';
+@import './components/size-label/size-label';
 
 /* todo: move styles to <l-quantity-selector>: DEZ-766 */
 .l-quantity-selector-component {
