+++ src/app/core/index.ts
@@ -1,2 +1,3 @@
 export * from './core.module';
 export * from './random-number.service';
+export * from './router-wrapper';
