+++ src/app/main.ts
@@ -1,3 +1,5 @@
+import '../app.ng1/app.module';
+
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 import { AppModule } from './app.module';
 import { UpgradeModule } from '@angular/upgrade/static';
