--- src/app/ng1/index.ts
@@ -1,2 +1,0 @@
-export * from './ng1.module';
-export * from './router-wrapper';
