--- src/app/ng1/ng1.module.ts
@@ -1,8 +1,0 @@
-import { NgModule } from '@angular/core';
-
-import { RouterWrapper } from './router-wrapper';
-
-@NgModule({
-  providers: [RouterWrapper]
-})
-export class Ng1Module {}
